feat: reorganise folder structure
