+++ src/components/organisms/pupil/index.ts
@@ -1,35 +1,3 @@
-export * from "./OakQuizCheckBox";
-export * from "./OakQuizTextInput";
-export * from "./OakQuizRadioButton";
-export * from "./OakQuizHint";
-export * from "./OakLessonBottomNav";
-export * from "./OakLessonTopNav";
-export * from "./OakQuizFeedback";
-export * from "./OakQuizCounter";
-export * from "./OakLessonLayout";
-export * from "./OakHintButton";
-export * from "./OakLessonNavItem";
-export * from "./OakLessonReviewItem";
-export * from "./OakLessonVideoTranscript";
-export * from "./OakQuizOrder";
-export * from "./OakQuizMatch";
-export * from "./OakPupilJourneyLayout";
-export * from "./Oakinfo";
-export * from "./OakinfoButton";
-export * from "./OakPupilJourneyListItem";
-export * from "./OakPupilJourneyYearButton";
-export * from "./OakPupilJourneySubjectButton";
-export * from "./OakPupilJourneyListItemSubheading";
-export * from "./OakPupilJourneyList";
-export * from "./OakPupilJourneyHeader";
-export * from "./OakPupilJourneyListCounter";
-export * from "./OakPupilJourneyProgrammeOptions";
-export * from "./OakPupilJourneyContentGuidance";
-export * from "./OakPupilJourneyOptionalityItem";
-export * from "./OakPupilJourneyOptionalityButton";
-export * from "./OakPupilJourneyUnitsFilter";
-export * from "./OakLessonReviewIntroVideo";
-export * from "./OakLessonReviewQuiz";
 export * from "./browse";
 export * from "./lesson";
 export * from "./quiz";
