+++ src/components/organisms/pupil/lesson/index.ts
@@ -1,6 +1,8 @@
 export * from "./OakLessonBottomNav";
 export * from "./OakLessonLayout";
 export * from "./OakLessonNavItem";
+export * from "./OakLessonReviewIntroVideo";
 export * from "./OakLessonReviewItem";
+export * from "./OakLessonReviewQuiz";
 export * from "./OakLessonTopNav";
 export * from "./OakLessonVideoTranscript";
