+++ src/components/organisms/teacher/index.ts
@@ -1,6 +1,7 @@
 export * from "./OakSearchFilterCheckBox";
 export * from "./OakTertiaryOLNav";
 export * from "./OakUnitListItem";
+export * from "./OakUnitListOptionalityItem";
+export * from "./OakUnitListOptionalityItemCard";
 export * from "./OakUnitsContainer";
 export * from "./OakUnitsHeader";
-export * from "./OakUnitListOptionalityItem";
